Rename CovidModule to NetworkModule
