+++ app/src/main/java/hu/bme/aut/pribelszki/covidio/di/AppComponent.kt
@@ -3,7 +3,7 @@ package hu.bme.aut.pribelszki.covidio.di
 import co.zsmb.rainbowcake.dagger.RainbowCakeComponent
 import co.zsmb.rainbowcake.dagger.RainbowCakeModule
 import dagger.Component
-import hu.bme.aut.pribelszki.covidio.network.CovidModule
+import hu.bme.aut.pribelszki.covidio.network.NetworkModule
 import hu.bme.aut.pribelszki.covidio.room.RoomModule
 import javax.inject.Singleton
 
@@ -13,7 +13,7 @@ import javax.inject.Singleton
         RainbowCakeModule::class,
         ViewModelModule::class,
         ApplicationModule::class,
-        CovidModule::class,
+        NetworkModule::class,
         RoomModule::class
     ]
 )
