+++ app/src/main/java/hu/bme/aut/pribelszki/covidio/network/NetworkModule.kt
